+++ src/dflow/common.py
@@ -12,7 +12,7 @@ class S3Artifact(V1alpha1S3Artifact):
     Args:
         key: key of the s3 artifact
     """
-    def __init__(self, path_list=None, *args, **kwargs):
+    def __init__(self, path_list=".", *args, **kwargs):
         config = Configuration()
         config.client_side_validation = False
         super().__init__(local_vars_configuration=config, *args, **kwargs)