+++ src/dflow/config.py
@@ -5,5 +5,6 @@ config = {
     "k8s_api_server": None,
     "private_key_host_path": "/home/docker/.ssh",
     "save_path_as_parameter": True,
-    "catalog_file_name": ".dflow"
+    "catalog_file_name": ".dflow",
+    "archive_mode": "tar",
 }